Fix rate limit handling

+++ NetCord/Rest/RateLimitedException.cs
@@ -3,9 +3,9 @@
 public class RateLimitedException : Exception
 {
     public bool Global { get; }
-    public int Reset { get; }
+    public long Reset { get; }
 
-    public RateLimitedException(int reset, bool global) : base("Rate limit triggered.")
+    public RateLimitedException(long reset, bool global) : base("Rate limit triggered.")
     {
         Reset = reset;
         Global = global;
